* Fix skipping of photos based on content type * Refactor provisioning * Add more tests to verify processes
bumping version
