+++ lib/socialcast.rb
@@ -2,8 +2,10 @@ require 'yaml'
 require 'fileutils'
 
 require_relative 'socialcast/command_line/ldap_connector'
-require_relative 'socialcast/command_line/provision'
+require_relative 'socialcast/command_line/provisioner'
 require_relative 'socialcast/command_line/authenticate'
+require_relative 'socialcast/command_line/provision_user'
+require_relative 'socialcast/command_line/provision_photo'
 require_relative 'socialcast/command_line/message'
 require_relative 'socialcast/command_line/cli'
 require_relative 'socialcast/command_line/version'
