+++ lib/socialcast/command_line/version.rb
@@ -1,5 +1,5 @@
 module Socialcast
   module CommandLine
-    VERSION = "1.3.7"
+    VERSION = "1.3.8"
   end
 end
